%[ER: 220]% Fixed 64-bit bugs.

+++ LCS/Common/src/shmem/dlmalloc.c
@@ -4955,7 +4955,7 @@ static void *sbrk (long size) {
     static region_list_entry *g_last;
     void *result = (void *) MORECORE_FAILURE;
 #ifdef TRACE
-    printf ("sbrk %d\n", size);
+    printf ("sbrk %ld\n", size);
 #endif
 #if defined (USE_MALLOC_LOCK) && defined (NEEDED)
     /* Wait for spin lock */
@@ -5012,7 +5012,7 @@ static void *sbrk (long size) {
                         /* Assert postconditions */
                         assert ((unsigned) base_committed % g_pagesize == 0);
 #ifdef TRACE
-                        printf ("Commit %p %d\n", base_committed, remaining_commit_size);
+                        printf ("Commit %p %ld\n", base_committed, remaining_commit_size);
 #endif
                         /* Adjust the regions commit top */
                         g_last->top_committed = (char *) base_committed + remaining_commit_size;
@@ -5038,7 +5038,7 @@ static void *sbrk (long size) {
                             /* Assert postconditions */
                             assert ((unsigned) memory_info.BaseAddress % g_pagesize == 0);
 #ifdef TRACE
-                            printf ("Query %p %d %s\n", memory_info.BaseAddress, memory_info.RegionSize, 
+                            printf ("Query %p %ld %s\n", memory_info.BaseAddress, (long) memory_info.RegionSize, 
                                     memory_info.State == MEM_FREE ? "FREE": 
                                     (memory_info.State == MEM_RESERVE ? "RESERVED":
                                      (memory_info.State == MEM_COMMIT ? "COMMITTED": "?")));
@@ -5082,7 +5082,7 @@ static void *sbrk (long size) {
                     /* Assert postconditions */
                     assert ((unsigned) base_reserved % g_regionsize == 0);
 #ifdef TRACE
-                    printf ("Reserve %p %d\n", base_reserved, reserve_size);
+                    printf ("Reserve %p %ld\n", base_reserved, reserve_size);
 #endif
                     /* Did we get contiguous memory? */
                     if (contiguous) {
@@ -5120,7 +5120,7 @@ static void *sbrk (long size) {
                 /* Assert postconditions */
                 assert ((unsigned) base_committed % g_pagesize == 0);
 #ifdef TRACE
-                printf ("Commit %p %d\n", base_committed, commit_size);
+                printf ("Commit %p %ld\n", base_committed, commit_size);
 #endif
                 /* Adjust the regions commit top */
                 g_last->top_committed = (char *) base_committed + commit_size;
@@ -5148,7 +5148,7 @@ static void *sbrk (long size) {
                 if (! rc)
                     goto sbrk_exit;
 #ifdef TRACE
-                printf ("Release %p %d\n", base_reserved, release_size);
+                printf ("Release %p %ld\n", base_reserved, release_size);
 #endif
             }
             /* Adjust deallocation size */
@@ -5174,7 +5174,7 @@ static void *sbrk (long size) {
                     if (! rc)
                         goto sbrk_exit;
 #ifdef TRACE
-                    printf ("Decommit %p %d\n", base_committed, decommit_size);
+                    printf ("Decommit %p %ld\n", base_committed, decommit_size);
 #endif
                 }
                 /* Adjust deallocation size and regions commit and allocate top */
@@ -5217,7 +5217,7 @@ static void *mmap (void *ptr, long size, long prot, long type, long handle, long
     static long g_pagesize;
     static long g_regionsize;
 #ifdef TRACE
-    printf ("mmap %d\n", size);
+    printf ("mmap %ld\n", size);
 #endif
 #if defined (USE_MALLOC_LOCK) && defined (NEEDED)
     /* Wait for spin lock */
@@ -5241,7 +5241,7 @@ static void *mmap (void *ptr, long size, long prot, long type, long handle, long
     /* Assert postconditions */
     assert ((unsigned) ptr % g_regionsize == 0);
 #ifdef TRACE
-    printf ("Commit %p %d\n", ptr, size);
+    printf ("Commit %p %ld\n", ptr, size);
 #endif
 mmap_exit:
 #if defined (USE_MALLOC_LOCK) && defined (NEEDED)
@@ -5257,7 +5257,7 @@ static long munmap (void *ptr, long size) {
     static long g_regionsize;
     int rc = MUNMAP_FAILURE;
 #ifdef TRACE
-    printf ("munmap %p %d\n", ptr, size);
+    printf ("munmap %p %ld\n", ptr, size);
 #endif
 #if defined (USE_MALLOC_LOCK) && defined (NEEDED)
     /* Wait for spin lock */
@@ -5277,7 +5277,7 @@ static long munmap (void *ptr, long size) {
         goto munmap_exit;
     rc = 0;
 #ifdef TRACE
-    printf ("Release %p %d\n", ptr, size);
+    printf ("Release %p %ld\n", ptr, size);
 #endif
 munmap_exit:
 #if defined (USE_MALLOC_LOCK) && defined (NEEDED)
